Live (#80)
* Trying to fix udevrules for vectornav
Changed velodyne orientation to offset it 180deg.
This will probably have an impact in the simulation in gazebo.

Adjusted the launch script to stop can bus on exit.

Disabled odrive id 24 rr_track (rear right track).

Disabled slam_3d to simplify testing.

* ROBOT MOVEDDDDD
Changed the position of the sensors in the URDF (they wer showed as if they were in the front instead of the back)
Modified the inputs for slam

---------

Co-authored-by: nathan-gt <nathan.gt.protic@gmail.com>