Insta360 x4 featuring pan zoom (#81)
* add insta360 launch file and udev rule creation

* rename config for the two camera modes

* add usb-cam to package.xml

* add image splitting for operator

* only splitting for requested angle

* add zoom

* add back camera and rename package

* Live (#80)

* Trying to fix udevrules for vectornav
Changed velodyne orientation to offset it 180deg.
This will probably have an impact in the simulation in gazebo.

Adjusted the launch script to stop can bus on exit.

Disabled odrive id 24 rr_track (rear right track).

Disabled slam_3d to simplify testing.

* ROBOT MOVEDDDDD
Changed the position of the sensors in the URDF (they wer showed as if they were in the front instead of the back)
Modified the inputs for slam

---------

Co-authored-by: nathan-gt <nathan.gt.protic@gmail.com>

* adding vertical control insta360 and fixing back view stitching

* fix package ref in config

* select insta nodes in launch

* add forked usb cam repo

* fix last comit

* Revert "add forked usb cam repo"

This reverts commit b89f7185882859df62b0b0f9af3caf10b56686bc.

* fix math, need TI

---------

Co-authored-by: nathan-gt <nathan.gt.protic@gmail.com>